Postgres sequences are 64-bit
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/pathsoft/genedb-ng/trunk@10727 ee4ac58c-ac51-4696-9907-e4b3aa274f04

+++ genedb-db/src/org/genedb/db/audit/HibernateChangeSet.java
@@ -18,9 +18,9 @@ public class HibernateChangeSet implements ChangeSet {
 
     private Session session;
     private String key;
-    private int ceilingAuditId;
+    private long ceilingAuditId;
 
-    public HibernateChangeSet(Session session, String key, int ceilingAuditId) {
+    public HibernateChangeSet(Session session, String key, long ceilingAuditId) {
         this.session = session;
         this.ceilingAuditId = ceilingAuditId;
         this.key = key;
@@ -95,7 +95,7 @@ public class HibernateChangeSet implements ChangeSet {
     public void commit() throws SQLException {
         int n = session.createSQLQuery(
             "update audit.checkpoint set audit_id = :ceiling where key = :key"
-        ).setInteger("ceiling", ceilingAuditId)
+        ).setLong("ceiling", ceilingAuditId)
         .setString("key", key)
         .executeUpdate();
 
@@ -103,7 +103,7 @@ public class HibernateChangeSet implements ChangeSet {
             logger.info(String.format("No existing audit.checkpoint record for key '%s'", key));
             session.createSQLQuery(
                 "insert into audit.checkpoint (key, audit_id) values (:key, :ceiling)"
-            ).setInteger("ceiling", ceilingAuditId)
+            ).setLong("ceiling", ceilingAuditId)
             .setString("key", key)
             .executeUpdate();
         }
